Spring security not quite working
The httpbasic *was* working but subsequently broken, and I have not yet
figured out how to get the form to invoke the servlet I want it to.
Added logout code to the vaadin7-demo, which does odd things when using
httpbasic.

So not much of this is actually working yet.

However the login form does popup when configured.